Move checkbox toggle to separate module. Remove unnecessary imports. Add animation to refresh icon.
