+++ src/modules/clearChecked.js
@@ -1,13 +1,12 @@
 import { getTasks, setTasks } from './storage.js';
-import render from './render.js';
 
 function clearCompletedTasks(tasks) {
   tasks = getTasks();
-    tasks = tasks.filter((task) => !task.completed);
-    tasks.forEach((task, index) => {
-      task.index = index + 1;
-    });
-    setTasks(tasks);
+  tasks = tasks.filter((task) => !task.completed);
+  tasks.forEach((task, index) => {
+    task.index = index + 1;
+  });
+  setTasks(tasks);
 }
 
 export default clearCompletedTasks;
